[finance_report_vision] refactor: split common/ci into runtime + testing
Second PR of the 5-PR series dissolving common/ci, common/shell, and
common/ssot into their owning packages. common/ci/ was a standalone
14-file directory mixing CI-execution decisions (test-facing) with
deployment/environment orchestration (runtime-facing).

- runtime (7 files): check_toolchain_contract.py, ghcr_retention.py,
  release_coordinate.py, release_images.py, wait_post_merge_train_turn.py,
  production_infra_smoke.py, release_evidence.py -- deployment/release/
  environment machinery, matching runtime's existing "app<->external-world
  dependency boundary" charter (its own MANIFEST concept
  runtime_toolchain already cross-referenced check_toolchain_contract.py
  next to common/shell/common.sh before this PR).
- testing (4 files): change_classifier.py, check_pr_review_threads.py,
  wait_for_cheap_ci.py, github_workflow_timing_summary.py -- CI
  test-execution decisions and merge-gate/reporting concerns.

metrics_contract.py, migration_risk.py, workflow_contract.py stay in
common/ci/ for now -- they're package/contract-governance tools headed to
common/meta/extension/ in a later PR of this series, not this slice.

git mv preserves history. Fixed both import styles
(`from common.ci.X import Y` and `from common.ci import X`) across 11
tools/*.py wrappers and ~12 test files, plus doc/config path mentions in
MANIFEST.yaml, ci-cd.md, runtime-incident-response.md,
DELIVERY_ENGINE_RECOMMENDATIONS.md, common/coverage/policy.py's coverage
exclude pattern, and the .github/workflows/ci.yml coveragerc omit list.

Verified: check_package_contract, check_authority_reconcile,
check_ac_index, lint_doc_consistency, check_manifest,
check_workflow_contract, check_toolchain_contract,
check_ci_metrics_contract all pass; the full tests/tooling/ suite (1843
passed) is clean; no apps/backend/src/* or apps/backend/tests/* file
references any of the 11 moved modules (no Docker-image-safety concern).
