+++ common/runtime/release_images.py
