+++ docs/ssot/ci-cd.md
@@ -75,7 +75,7 @@ main-only: unified-coverage ─→ unified-coverage-baseline-pr (not required by
 17. **Deny-list coverage scope**: Coverage scope is deny-list based within each governed source root. CI recursively expects every eligible source file in backend, frontend, common, and tools LCOV unless `common/coverage/policy.py` explicitly excludes it. New source roots fail the metrics contract until added to the policy and report pipeline.
 18. **Contract checks run in both `lint` and `tooling-coverage` by design**: Each SSOT/contract script the `lint` job runs (`check_manifest`, `check_workflow_contract`, `check_governance_exceptions`, `generate_api_reference --check`, `check_ci_metrics_contract`, `check_ac_index`, `audit_router_contracts`, `lint_doc_consistency`, `generate_epic_status`, etc.) is **also** asserted by a `tests/tooling/` test executed in the `tooling-coverage` job (e.g. `test_check_manifest.py`, `test_workflow_contract.py`), several of which assert that the real repository passes. The two jobs serve different masters and are not redundant: `lint` is the standalone fast-fail lane (no test environment, surfaces in ~1 min, Key CI Property 1), while `tooling-coverage` executes those contracts as tests so `common/`+`tools/` get line coverage. The consequence is intentional but worth knowing: **a single contract drift fails both `lint` and `tooling-coverage`** — fix the one source, not each job. The `pre-push` tier (see [Local Guardrail Tiers](#local-guardrail-tiers-and-latency-budget)) mirrors the cheapest of these so the drift is usually caught before either CI job runs.
 19. **`finish` reporting is non-gating**: The `finish` job's merge authority is its `Check job status` step alone. Its three step-summary writers (coverage-effect table, CI timing, coverage-gate narrative) are reporting only and all carry `continue-on-error`, so a summary write can never decide merge.
-20. **`container-images` — narrow on PR, always builds the immutable `:<sha>` on main**: The classifier (`common/ci/change_classifier.py`) emits `image_build_required`, true only when a Docker build input changed — Dockerfiles, `.dockerignore` (changes what is sent in the build context), dependency manifests/locks (`uv.lock`, `package-lock.json`, `pyproject.toml`, `package.json`), build config (`next.config.mjs`, `tsconfig.json`, `postcss`/`tailwind` config), or the backend entrypoint/build scripts (`apps/backend/scripts/`). **On a PR** that signal narrows the dry-run: a pure app-source change is validated by `frontend-build` (`tsc --noEmit` + `next build`) and the backend test jobs, so the PR does **not** also dry-run a fresh image build (no push happens on PRs anyway). **On a `main`/`release` push (and manual dispatch) `container-images` ALWAYS builds and pushes the `:<sha>` image regardless of `image_build_required`** — the Dockerfiles `COPY` app source into the image, so every commit's image content differs, and promote-not-rebuild (the `deploy_v2` release path) requires an immutable `:<sha>` artifact for *any* releasable commit; gating the main build on build-context alone would leave source-only / submodule-only commits with no `:<sha>` to promote. `container-images` therefore gates on `(push to main|release) || workflow_dispatch || (pr_required && image_build_required)`; `finish` treats a skipped `container-images` (PR-only) as a pass (not a gap). An unknown diff fails closed and rebuilds. This keeps PRs fast while guaranteeing every main commit is releasable.
+20. **`container-images` — narrow on PR, always builds the immutable `:<sha>` on main**: The classifier (`common/testing/change_classifier.py`) emits `image_build_required`, true only when a Docker build input changed — Dockerfiles, `.dockerignore` (changes what is sent in the build context), dependency manifests/locks (`uv.lock`, `package-lock.json`, `pyproject.toml`, `package.json`), build config (`next.config.mjs`, `tsconfig.json`, `postcss`/`tailwind` config), or the backend entrypoint/build scripts (`apps/backend/scripts/`). **On a PR** that signal narrows the dry-run: a pure app-source change is validated by `frontend-build` (`tsc --noEmit` + `next build`) and the backend test jobs, so the PR does **not** also dry-run a fresh image build (no push happens on PRs anyway). **On a `main`/`release` push (and manual dispatch) `container-images` ALWAYS builds and pushes the `:<sha>` image regardless of `image_build_required`** — the Dockerfiles `COPY` app source into the image, so every commit's image content differs, and promote-not-rebuild (the `deploy_v2` release path) requires an immutable `:<sha>` artifact for *any* releasable commit; gating the main build on build-context alone would leave source-only / submodule-only commits with no `:<sha>` to promote. `container-images` therefore gates on `(push to main|release) || workflow_dispatch || (pr_required && image_build_required)`; `finish` treats a skipped `container-images` (PR-only) as a pass (not a gap). An unknown diff fails closed and rebuilds. This keeps PRs fast while guaranteeing every main commit is releasable.
 
 ### Env x Stage Contract
 
@@ -363,7 +363,7 @@ Tier 1, Tier 2, and Tier 3 evidence.
 ## PR Review Thread Merge Gate
 
 `tools/check_pr_review_threads.py` (logic in
-`common/ci/check_pr_review_threads.py`) is a merge-time gate that blocks a PR
+`common/testing/check_pr_review_threads.py`) is a merge-time gate that blocks a PR
 while a high-severity review thread is still open. It runs as a step in the
 `lint` job of `.github/workflows/ci.yml` on `pull_request` events, reading the
 PR's review threads through the GitHub GraphQL API (`gh api graphql`). It is
