+++ tests/tooling/test_toolchain_contract.py
@@ -9,7 +9,7 @@ import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-from common.ci import check_toolchain_contract as contract  # noqa: E402
+from common.runtime import check_toolchain_contract as contract  # noqa: E402
 
 
 ROOT = Path(__file__).resolve().parents[2]
