+++ tools/ci_change_classifier.py
@@ -10,7 +10,7 @@ ROOT_DIR = Path(__file__).resolve().parents[1]
 if str(ROOT_DIR) not in sys.path:
     sys.path.insert(0, str(ROOT_DIR))
 
-from common.ci.change_classifier import main  # noqa: E402
+from common.testing.change_classifier import main  # noqa: E402
 
 
 if __name__ == "__main__":  # pragma: no cover
